Move tap annotation to `viz/pkg/labels`
