+++ cli/cmd/inject_test.go
@@ -602,17 +602,17 @@ func TestProxyConfigurationAnnotations(t *testing.T) {
 	values.GetGlobal().Proxy.WaitBeforeExitSeconds = 10
 
 	expectedOverrides := map[string]string{
-		k8s.ProxyIgnoreInboundPortsAnnotation:     "8500-8505",
-		k8s.ProxyIgnoreOutboundPortsAnnotation:    "3306",
-		k8s.ProxyAdminPortAnnotation:              "1234",
-		k8s.ProxyControlPortAnnotation:            "4191",
-		k8s.ProxyInboundPortAnnotation:            "4144",
-		k8s.ProxyOutboundPortAnnotation:           "4141",
-		k8s.ProxyUIDAnnotation:                    "999",
-		k8s.ProxyLogLevelAnnotation:               "debug",
-		k8s.ProxyLogFormatAnnotation:              "cool",
-		k8s.ProxyDisableIdentityAnnotation:        "true",
-		k8s.ProxyDisableTapAnnotation:             "true",
+		k8s.ProxyIgnoreInboundPortsAnnotation:  "8500-8505",
+		k8s.ProxyIgnoreOutboundPortsAnnotation: "3306",
+		k8s.ProxyAdminPortAnnotation:           "1234",
+		k8s.ProxyControlPortAnnotation:         "4191",
+		k8s.ProxyInboundPortAnnotation:         "4144",
+		k8s.ProxyOutboundPortAnnotation:        "4141",
+		k8s.ProxyUIDAnnotation:                 "999",
+		k8s.ProxyLogLevelAnnotation:            "debug",
+		k8s.ProxyLogFormatAnnotation:           "cool",
+		k8s.ProxyDisableIdentityAnnotation:     "true",
+
 		k8s.ProxyEnableExternalProfilesAnnotation: "true",
 		k8s.ProxyCPURequestAnnotation:             "10m",
 		k8s.ProxyCPULimitAnnotation:               "100m",
