+++ testutil/inject_validator.go
@@ -25,7 +25,6 @@ type InjectValidator struct {
 	AutoInject             bool
 	AdminPort              int
 	ControlPort            int
-	DisableTap             bool
 	EnableDebug            bool
 	EnableExternalProfiles bool
 	ImagePullPolicy        string
@@ -133,12 +132,6 @@ func (iv *InjectValidator) validateProxyContainer(pod *v1.PodSpec) error {
 
 	if iv.ControlPort != 0 {
 		if err := iv.validateEnvVar(proxyContainer, "LINKERD2_PROXY_CONTROL_LISTEN_ADDR", fmt.Sprintf("0.0.0.0:%d", iv.ControlPort)); err != nil {
-			return err
-		}
-	}
-
-	if iv.DisableTap {
-		if err := iv.validateNoEnvVar(proxyContainer, "LINKERD2_PROXY_TAP_SVC"); err != nil {
 			return err
 		}
 	}
@@ -420,10 +413,6 @@ func (iv *InjectValidator) GetFlagsAndAnnotations() ([]string, map[string]string
 	if iv.DisableIdentity {
 		annotations[k8s.IdentityModeDisabled] = enabled
 		flags = append(flags, "--disable-identity")
-	}
-
-	if iv.DisableTap {
-		annotations[k8s.ProxyDisableTapAnnotation] = enabled
 	}
 
 	if iv.EnableDebug {
