+++ viz/pkg/labels/labels.go
@@ -13,12 +13,34 @@ const (
 	// VizTapEnabled is set by the tap-injector component when tap has been
 	// enabled on a pod.
 	VizTapEnabled = VizAnnotationsPrefix + "/tap-enabled"
+
+	// VizTapDisabled can be used to disable tap on the injected proxy.
+	VizTapDisabled = VizAnnotationsPrefix + "/disable-tap"
 )
 
 // IsTapEnabled returns true if a pod has an annotation indicating that tap
 // is enabled.
 func IsTapEnabled(pod *corev1.Pod) bool {
 	valStr := pod.GetAnnotations()[VizTapEnabled]
+	if valStr != "" {
+		valBool, err := strconv.ParseBool(valStr)
+		if err == nil && valBool {
+			return true
+		}
+	}
+	return false
+}
+
+// IsTapDisabled returns true if a namespace or pod has an annotation for
+// explicitly disabling tap
+func IsTapDisabled(obj interface{}) bool {
+	var valStr string
+	switch resource := obj.(type) {
+	case *corev1.Pod:
+		valStr = resource.GetAnnotations()[VizTapDisabled]
+	case *corev1.Namespace:
+		valStr = resource.GetAnnotations()[VizTapDisabled]
+	}
 	if valStr != "" {
 		valBool, err := strconv.ParseBool(valStr)
 		if err == nil && valBool {
